refactor: 파일 정리 및 기본 값 제거
