+++ index.html
@@ -20,7 +20,7 @@
         <form id="purchase-form">
           <label for="price">구입할 금액을 입력해주세요.</label>
           <div class="purchase-input">
-            <input value="10000" type="number" id="price" placeholder="금액" />
+            <input type="number" id="price" placeholder="금액" />
             <button class="typo-caption">구입</button>
           </div>
           <p class="error-message typo-caption" id="price-error"></p>
@@ -35,18 +35,18 @@
             <div class="winning-number-group">
               <label>당첨 번호</label>
               <div class="winning-number-boxes">
-                <input type="text" id="winning-1" value="3" class="number-box" maxlength="2" />
-                <input type="text" id="winning-2" value="5" class="number-box" maxlength="2" />
-                <input type="text" id="winning-3" value="7" class="number-box" maxlength="2" />
-                <input type="text" id="winning-4" value="9" class="number-box" maxlength="2" />
-                <input type="text" id="winning-5" value="11" class="number-box" maxlength="2" />
-                <input type="text" id="winning-6" value="22" class="number-box" maxlength="2" />
+                <input type="text" id="winning-1" class="number-box" maxlength="2" />
+                <input type="text" id="winning-2" class="number-box" maxlength="2" />
+                <input type="text" id="winning-3" class="number-box" maxlength="2" />
+                <input type="text" id="winning-4" class="number-box" maxlength="2" />
+                <input type="text" id="winning-5" class="number-box" maxlength="2" />
+                <input type="text" id="winning-6" class="number-box" maxlength="2" />
               </div>
             </div>
 
             <div class="bonus-number-group">
               <label for="bonus">보너스 번호</label>
-              <input type="text" id="bonus" class="number-box bonus-box" value="23" maxlength="2" />
+              <input type="text" id="bonus" class="number-box bonus-box" maxlength="2" />
             </div>
           </div>
           <div class="spacing-4"></div>
@@ -103,6 +103,6 @@
       </div>
     </main>
     <footer><p class="typo-caption flex-center">Copyright 2023. woowacourse</p></footer>
-    <script src="/src/js/index.js" type="module"></script>
+    <script src="/src/step2-index.js" type="module"></script>
   </body>
 </html>
